Modify : 헤더 추가. 컴포넌트 분할.
