+++ src/Pages/CategoryPage/Components/index.jsx
@@ -1,0 +1,3 @@
+import Body from './Body';
+
+export default Body;
